feat(auth): add authorization gates and middleware
Define authorization gates for updating events and deleting
attendees. Add middleware to AttendeeController to require
authentication. Implement authorization checks in EventController
and AttendeeController destroy methods.

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,12 +8,22 @@ use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
     use CanLoadRelationships;
 
     private array $relations = ["user"];
+
+
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show', 'update']),
+        ];
+    }
 
     /**
      * Display a listing of the resource.
@@ -60,6 +70,9 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event, Attendee $attendee)
     {
+        // $this->authorize('delete-attendee', [$event, $attendee]);
+        Gate::authorize('delete-attendee', [$event, $attendee]);
+
         $attendee->delete();
 
         return response()->noContent();
